api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.1.2",
    "flask==3.0.2",
    "flask-cors==4.0.0",
    "flask-jwt-extended==4.6.0",
    "flask-sqlalchemy==3.1.1",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.25",
    "typing-extensions>=4.8.0",
]
